FROEST-FIRE-DETECTION-main/forest-fire-detection/requirements.txt
numpy
pillow
flask
flask-cors
opencv-python
matplotlib
scikit-learn
pytest
requests
